moving the download calendar option to assets/calendar.js but I left a method in store/planner to make calling it easier
